feat: refactored migration structure in all pallets
