+++ pallets/delegation/Cargo.toml
@@ -71,6 +71,6 @@ std = [
 ]
 try-runtime = [
   "frame-support/try-runtime",
-  "kilt-traits/try-runtime",
   "kilt-primitives/try-runtime",
+  "kilt-traits/try-runtime",
 ]
