+++ pallets/did/Cargo.toml
@@ -56,6 +56,7 @@ std = [
   "frame-support/std",
   "frame-system/std",
   "kilt-primitives/std",
+  "kilt-traits/std",
   "hex/std",
   "log/std",
   "sp-core/std",
@@ -66,5 +67,6 @@ std = [
 ]
 try-runtime = [
   "frame-support/try-runtime",
+  "kilt-traits/try-runtime",
   "kilt-primitives/try-runtime",
 ]
