+++ pallets/did/src/deprecated/v2.rs
@@ -19,13 +19,20 @@
 use codec::{Decode, Encode};
 use kilt_primitives::Hash;
 
-pub(crate) use super::*;
-use crate::*;
+use crate::{*, deprecated::Url};
+use sp_std::vec::Vec;
 
 #[derive(Clone, Decode, Encode, PartialEq, Eq)]
 pub(crate) enum ContentType {
 	ApplicationJson,
 	ApplicationJsonLd,
+}
+
+#[derive(Clone, Decode, Encode, PartialEq)]
+pub(crate) struct ServiceEndpoints {
+	pub content_hash: Hash,
+	pub urls: Vec<Url>,
+	pub content_type: ContentType,
 }
 
 #[derive(Clone, Decode, Encode, PartialEq)]
@@ -37,13 +44,6 @@ pub struct DidDetails<T: Config> {
 	pub(crate) public_keys: DidPublicKeyMap<T>,
 	pub(crate) service_endpoints: Option<ServiceEndpoints>,
 	pub(crate) last_tx_counter: u64,
-}
-
-#[derive(Clone, Decode, Encode, PartialEq)]
-pub(crate) struct ServiceEndpoints {
-	pub content_hash: Hash,
-	pub urls: Vec<Url>,
-	pub content_type: ContentType,
 }
 
 pub(crate) mod storage {
