+++ pallets/did/src/lib.rs
@@ -134,7 +134,7 @@ use frame_support::{
 use frame_system::ensure_signed;
 #[cfg(feature = "runtime-benchmarks")]
 use frame_system::RawOrigin;
-use kilt_primitives::StorageMigrator;
+use kilt_primitives::migrations::StorageMigrator;
 use sp_runtime::{traits::Saturating, SaturatedConversion};
 use sp_std::{boxed::Box, fmt::Debug, prelude::Clone};
